+++ interface/src/lib/components/admin/login.svelte
@@ -20,7 +20,7 @@
       localStorage.setItem("admin-password", pass);
       localStorage.setItem("admin-access-token", accessToken);
       loginLoading.set(false);
-      goto("/manage/dashboard");
+      goto("/manage/settings");
       return true;
     }
     submitted.set(true);